+++ core/src/test/java/com/bc/fiduceo/db/QueryParameterTest.java
@@ -26,6 +26,7 @@ import org.junit.Test;
 import java.util.Date;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 
 public class QueryParameterTest {
 
@@ -34,6 +35,18 @@ public class QueryParameterTest {
     @Before
     public void setUp() {
         parameter = new QueryParameter();
+    }
+
+    @Test
+    public void testConstruction() {
+        assertNull(parameter.getStartTime());
+        assertNull(parameter.getStopTime());
+        assertNull(parameter.getSensorName());
+        assertNull(parameter.getVersion());
+        assertNull(parameter.getPath());
+
+        assertEquals(-1, parameter.getPageSize());
+        assertEquals(-1, parameter.getOffset());
     }
 
     @Test
@@ -86,5 +99,29 @@ public class QueryParameterTest {
 
         parameter.setPath(path_2);
         assertEquals(path_2, parameter.getPath());
+    }
+
+    @Test
+    public void testSetGetPageSize() {
+        final int size_1= 58;
+        final int size_2= 108;
+
+        parameter.setPageSize(size_1);
+        assertEquals(size_1, parameter.getPageSize());
+
+        parameter.setPageSize(size_2);
+        assertEquals(size_2, parameter.getPageSize());
+    }
+
+    @Test
+    public void testSetGetOffset() {
+        final int offset_1= 34567;
+        final int offset_2= 100099746;
+
+        parameter.setOffset(offset_1);
+        assertEquals(offset_1, parameter.getOffset());
+
+        parameter.setOffset(offset_2);
+        assertEquals(offset_2, parameter.getOffset());
     }
 }