+++ CHANGES.md
@@ -3,6 +3,7 @@
 * added support for FIDUCEO AVHRR FCDR products
 * updated scripts to new JASMIN file system
 * added support for GRUAN reference files
+* added optional paging and offfset to database drivers
 
 ### Updates from version 1.4.0 to 1.4.1
 * migrated to asynchronous PMonitor